Trying tor fix compiler issue

+++ Sources/Numerical/SpecialFunctions/Beta.swift
@@ -237,7 +237,7 @@ public func invBetaReg(p: Probability, a: Double, b: Double) -> Double {
             let λ = (yp^^2 - 3.0) / 6.0
             let h = 2 / (1 / (2 * a - 1) + 1 / (2 * b - 1))
             let w = yp * sqrt(h + λ) / h - (1 / (2 * b - 1) - 1 / (2 * a - 1)) * (λ + 5 / 6 - 2 / (3 * h))
-            return a / (a + b * exp(2 * w))
+            return Double(a / (a + b * exp(2 * w)))
         // At least one of a and b < 1. Use NR approximation
         case (_,_):
             let lna = log(a / (a + b))
@@ -246,7 +246,7 @@ public func invBetaReg(p: Probability, a: Double, b: Double) -> Double {
             let u = exp(b * lnb) / b
             let w = t + u
             if p.p < t / w { return pow(a * w * p.p, 1 / a) }
-            return 1 - pow(b * w * p.q, 1 / b)
+            return Double(1 - pow(b * w * p.q, 1 / b))
         }
     }()
